requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
Django==1.11.3
djangorestframework==3.6.3
dlib==19.4.0
face-recognition==0.2.0
face-recognition-models==0.1.3
future==0.16.0
idna==2.5
numpy==1.13.0
olefile==0.44
Pillow==4.2.0
pytz==2017.2
requests==2.18.1
scipy==0.19.1
urllib3==1.21.1
